Modernize vector classes with base::ssize and fix static analyzer issues
- Change all indexes, sizes, and counts to use base::ssize for
  consistent signed integer semantics
- Use C++20 requires clause on Delete(object) to disambiguate from
  Delete(index) when T is a pointer type and 0 is passed
- Fix BooleanVectorClass::Resize to memset new bytes instead of
  reading uninitialized memory via Get_Bit
- Add bounds checks in loaddlg.cpp before accessing Files vector
- Fix delete vs delete[] mismatch in wspudp.cpp for array allocations

+++ ra/list.cpp
@@ -446,7 +446,7 @@ char const *ListClass::Get_Item(size_t index) const {
   if (List.Count() == 0) {
     return nullptr;
   }
-  index = std::min(index, List.Count() - 1);
+  index = std::min(index, static_cast<size_t>(List.Count() - 1));
   return (List[index]);
 }
